requirements.txt
fastapi
uvicorn
torch
transformers
python-dotenv
streamlit>=1.23.0
altair==4.2.2